refact: split mysql code
